+++ libs/notification/src/lib/invitation/+state/invitation.model.ts
@@ -1,18 +1,4 @@
 import { InvitationBase } from './invitation.firestore';
-export { InvitationStatus } from './invitation.firestore';
+export { InvitationStatus, createInvitation } from './invitation.firestore';
 
 export type Invitation = InvitationBase<Date>;
-
-/** Create an Invitation */
-export function createInvitation(params: Partial<Invitation> = {}): Invitation {
-  return  {
-    id: '',
-    mode: null,
-    type: null,
-    docId: '',
-    status: 'pending',
-    date: new Date(),
-    ...params,
-  };
-}
-
