+++ libs/user/src/lib/user/+state/user.firestore.ts
@@ -1,4 +1,20 @@
-import { ImgRef } from "@blockframes/utils/image-uploader";
+import { ImgRef, createImgRef } from "@blockframes/utils/image-uploader";
+
+export interface User {
+  uid: string;
+  financing: {
+    rank: string
+  };
+  email: string;
+  firstName: string;
+  lastName: string;
+  phoneNumber: string;
+  position: string;
+  orgId: string;
+  avatar: ImgRef;
+  watermark: ImgRef;
+}
+
 
 /** A user interface with public informations */
 export interface PublicUser {
@@ -9,4 +25,16 @@ export interface PublicUser {
   firstName?: string;
   lastName?: string;
   orgId?: string;
+}
+
+export function createPublicUser(user: Partial<User> = {}) : PublicUser{
+  return {
+    uid: user.uid || '',
+    email: user.email || '',
+    avatar: createImgRef(user.avatar),
+    watermark: createImgRef(user.watermark),
+    firstName: user.firstName || '',
+    lastName: user.lastName || '',
+    orgId: user.orgId || ''
+  }
 }
